+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## 0.0.14
+* Fixed Dio API breaking change
+
+## 0.0.13
+* Updated dependencies
+* Notification won't init when showNotification is off
+
+## 0.0.12
+* Updated flutter local notification dependency version
+* Refactor
+
+## 0.0.11
+* Fixed iOS version issues (fixed by https://github.com/britannio Britannio Jarrett)
+
+## 0.0.10
+* Added stats feature
+* Added save feature
+* Added secured/not secured connection indicator in call list item
+* Query parameters feature (Dio only)
+* Fix for Uint8List SDK breaking change
+* Updated dependencies
+* Refactored code
+
+## 0.0.6
+* Fixed http/http package requests
+
 ## 0.0.5
 * Updated dependencies
 * Navigator key can be provided now from application (instead of using Alice's navigator key)